+++ src/components/CreateComment.tsx
@@ -137,15 +137,15 @@ const CreateComment: FC<Props> = ({ publication }) => {
     const isLoading = typedDataLoading || signLoading || writeLoading
 
   return (
-   
-      <form onSubmit={createComment} className="flex-shrink-0 flex p-5 gap-3 border-t">
+    <div className="absolute bottom-0 left-0 pb-6 px-2 md:px-10">
+      <form onSubmit={createComment} className="flex gap-4">
         <input
           type="text"
           value={comment}
           onChange={(e) =>
             setComment(e.target.value)}
-            className="bg-[#F1F1F2] rounded-md p-2 flex-grow text-sm outline-none placeholder:text-gray-500 border border-transparent focus:border-gray-300 transition"
-            placeholder='Add comment..'
+          className='bg-primary px-6 py-4 text-md font-medium border-2 w-[250px] md:w-[700px] lg:w-[350px] border-gray-100 focus:outline-none focus:border-2 focus:border-gray-300 flex-1 rounded-lg'
+          placeholder='Add comment..'
         />
         <button className="text-md text-gray-400 border-gray-100" onClick={ createComment} disabled={isLoading}>
           {isLoading ? 'Commenting...' 
@@ -153,7 +153,7 @@ const CreateComment: FC<Props> = ({ publication }) => {
           'Comment'}
         </button>
       </form>
-  
+    </div>
   )
 }
