+++ src/components/UI/ProfileVideos.tsx
@@ -6,13 +6,9 @@ import { useQuery } from "@apollo/client";
 import { useRouter } from 'next/router';
 import type { FC } from "react";
 import { sanitizeIpfsUrl } from "@/utils/sanitizeIpfsUrl";
-import { BsPlay } from "react-icons/bs";
 
 
   const ProfileVideos = () => {
-    const [isPlaying, setIsPlaying] = useState<boolean>(false);
-
-    const videoRef = useRef<HTMLVideoElement>(null);
     const router = useRouter();
     const { id } = router.query
 
@@ -34,46 +30,33 @@ import { BsPlay } from "react-icons/bs";
     const publications = data?.publications.items;
     console.log("DATA", data?.publications?.items);
 
-    const onVideoClick = () => {
-        if (isPlaying) {
-          videoRef?.current?.pause();
-          setIsPlaying(false);
-        } else {
-          videoRef?.current?.play();
-          setIsPlaying(true);
-        }
-      };
-
     return (
-  <div>
-        {publications?.length === 0 ? (
-            <p className="text-center">No videos yet</p>
-           ) : (
-            <div className="grid gap-2 mr-2 mt-2 lg:grid-cols-3 md:gap-y-8 gap-y-2 3xl:grid-cols-4 md:grid-cols-3 sm:grid-cols-2 xs:grid-col-1">
-            {publications?.map((pub) => (
-                <div key={pub.id}>
-                    <Link href={`/detail/${pub.id}`} key={pub.id}>
-                        <a className="block h-0 relative pb-[131%]">
-                        <video
-                        ref={videoRef}
-                        loop
-                        src={sanitizeIpfsUrl(pub.metadata.media[0].original.url)}
-                        className="absolute inset-0 h-full w-full object-cover rounded"
-                        
-
-
-                        /> 
-                        <BsPlay onClick={onVideoClick} className="absolute left-3 bottom-3 fill-white w-7 h-7" />
-                        </a>
-                     </Link>
-                    <p className="whitespace-nowrap overflow-hidden text-ellipsis">
-                    {pub.metadata.name}
-                    </p>
-                </div>
-             ))}
-             </div>
-         )}
-  </div>
-            )}
+      <div>
+      <div className='flex gap-10 mb-10 mt-10 border-b-2 border-gray-200 bg-white w-full'>
+      <p className='text-xl font-semibold cursor-pointer mt-2'>Videos</p>
+      </div>
+      {publications ? (
+      <div className='grid gap-1 mt-2 lg:grid-cols-3 md:gap-y-8 gap-y-2 3xl:grid-cols-4 md:grid-cols-3 sm:grid-cols-2 xs:grid-col-1'>
+      {publications?.map((pub) => (
+        <div key={pub.id} className='relative w-full h-60 bg-cover bg-center bg-no-repeat'>
+            <Link href={`/detail/${pub.id}`} key={pub.id}>
+            <video
+            loop
+            src={sanitizeIpfsUrl(pub.metadata.media[0].original.url)}
+            className='bg-gray-100 rounded-xl dark:bg-[#181818] aspect-w-16 aspect-h-9 cursor-pointer w-full h-full lg:w-full lg:h-full object-contain'
+          ></video> 
+          </Link>
+          <p>{pub.metadata.name}</p>
+        </div>
+      ))}
+      </div>
+      ) : (
+          <div>
+            <p>No videos yet</p>
+          </div>
+      )} 
+    </div>
+    );
+  };
   
   export default ProfileVideos;
