--- src/graphql/profiles/Following.graphql
@@ -1,15 +1,0 @@
-query Following($request: FollowingRequest!) {
-  following(request: $request) {
-    items {
-      profile {
-        ...ProfileFields
-        isFollowedByMe
-      }
-      totalAmountOfTimesFollowing
-    }
-    pageInfo {
-      next
-      totalCount
-    }
-  }
-}
