--- src/utils/clipboard.ts
@@ -1,23 +1,0 @@
-export const copyToClipboard = (text: string) => {
-    try {
-      if (navigator.clipboard && window.isSecureContext) {
-        return navigator.clipboard.writeText(text);
-      } else {
-        let textArea = document.createElement("textarea");
-        textArea.value = text;
-        textArea.style.position = "fixed";
-        textArea.style.left = "-999999px";
-        textArea.style.top = "-999999px";
-        document.body.appendChild(textArea);
-        textArea.focus();
-        textArea.select();
-        return new Promise((res, rej) => {
-          document.execCommand("copy") ? res(undefined) : rej();
-          textArea.remove();
-        });
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
